feat: 实现了lseek 并且编写了针对triple indirect block的测试

+++ kernel/file.h
@@ -5,7 +5,7 @@ struct file {
   char writable;
   struct pipe *pipe; // FD_PIPE
   struct inode *ip;  // FD_INODE and FD_DEVICE
-  uint off;          // FD_INODE
+  uint64 off;          // FD_INODE
   short major;       // FD_DEVICE
 };
 
@@ -25,7 +25,7 @@ struct inode {
   short major;
   short minor;
   short nlink;
-  uint size;
+  uint64 size;
   uint addrs[NDIRECT+3];
 };
